requirements.txt
# HIRL - 人机交互强化学习平台依赖列表

# 基础依赖
numpy>=1.21.0
pygame>=2.1.0
gymnasium>=0.28.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# 强化学习相关
stable-baselines3[extra]>=2.0.0
torch>=1.12.0
tensorboard>=2.10.0

# 实验跟踪和可视化
wandb>=0.15.0
matplotlib>=3.5.0

# 数据处理和保存
pandas>=2.0.0
h5py>=3.8.0
datasets>=2.10.0
huggingface-hub>=0.14.0

# 图像处理
opencv-python>=4.6.0
Pillow>=9.0.0

# 其他实用工具
tqdm>=4.64.0